+++ src/main/java/com/tsystems/ecare/app/services/RoleService.java
@@ -2,9 +2,12 @@ package com.tsystems.ecare.app.services;
 
 import com.tsystems.ecare.app.model.Role;
 
+/**
+ * Service operations interface for Role entity.
+ */
 public interface RoleService {
 
-    public Role getRoleById(Integer id);
+    public Role getRoleById(Long id);
 
     public Role getRoleByTitle(String title);
